Initial IRR function
This initial version of the IRR function looks ok for most common usage
but should be refined. I used here the bisection method as an algorithm
for function root finding. But the function is too stupid to guess that
no IRR could be find in certain conditions, etc. Also, there seem to be
better algorithms out there for root finding.